+++ pymc_extras/inference/INLA/inla.py
@@ -19,7 +19,6 @@ def fit_INLA(
     **sampler_kwargs,
 ) -> az.InferenceData:
     warnings.warn("Currently only valid for a nested normal model. WIP.", UserWarning)
-    # TODO ADD CHECK FOR NESTED NORMAL
 
     model = pm.modelcontext(model)
