+++ pymc_extras/inference/__init__.py
@@ -13,7 +13,7 @@
 #   limitations under the License.
 
 from pymc_extras.inference.fit import fit
-from pymc_extras.inference.inla import fit_INLA
+from pymc_extras.inference.INLA.inla import fit_INLA
 from pymc_extras.inference.laplace_approx.find_map import find_MAP
 from pymc_extras.inference.laplace_approx.laplace import fit_laplace
 from pymc_extras.inference.pathfinder.pathfinder import fit_pathfinder
